ai-processing/requirements.txt
# PrismWeave AI Processing Requirements
# Core dependencies for local LLM integration and document processing

# Ollama integration
ollama>=0.1.7

# Vector database and embeddings
chromadb>=0.4.15
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4

# Text processing and NLP
nltk>=3.8.1
spacy>=3.7.2
textstat>=0.7.3
langdetect>=1.0.9

# Document parsing and manipulation
python-frontmatter>=1.0.0
markdown>=3.5.1
beautifulsoup4>=4.12.2
markdownify>=0.11.6

# Data handling and storage
pandas>=2.1.0
numpy>=1.24.0
sqlite3  # Built-in Python module
sqlalchemy>=2.0.0
pydantic>=2.4.0

# API and CLI tools
fastapi>=0.104.0
uvicorn>=0.24.0
click>=8.1.7
rich>=13.6.0
tqdm>=4.66.0

# Graph and relationship analysis
networkx>=3.1
scikit-learn>=1.3.0

# Configuration and utilities
pyyaml>=6.0.1
python-dotenv>=1.0.0
loguru>=0.7.2

# Performance monitoring
psutil>=5.9.0
memory-profiler>=0.61.0

# Testing and development
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-mock>=3.11.0
